fix(bios): replace the use with lazy_static with static Mutex.
I don't exactly know why I cannot use lazy_static in bios.

When running simulation, the program jump to spin-0.5.2/src/once.rs:103 and then read a wrong number to status, causing an `unreachable` panic. And after replacing it with a ugly implementation with Mutex and wrapper, it works.

The implementation of uart still need to be changed.